+++ client/karma.conf.js
@@ -14,7 +14,7 @@
 
             // frameworks to use
             // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-            frameworks: ['mocha', 'chai-as-promised', 'chai', 'sinon'],
+            frameworks: ['mocha', 'chai-as-promised', 'chai', 'sinon', 'sinon-chai'],
 
             // list of files / patterns to load in the browser
             files: [