+++ client/src/main/js/controller/AgendaController.js
@@ -22,9 +22,13 @@
              * @param reserva Reserva clicado.
              * @return {Promise} Promise do modal de visualização da reserva
              */
-            this.clickReserva = function (reserva) {
+            this.visualizarReserva = function (reserva) {
                 const reservaObj = new Reserva(reserva);
-                return ModalService.verReserva(reservaObj);
+                return ModalService.verReserva(reservaObj, self.local).then(info => {
+                    uiCalendarConfig.calendars.calendario.fullCalendar('removeEvents');
+                    uiCalendarConfig.calendars.calendario.fullCalendar('addEventSource', _.first(self.reservasSource));
+                    return info;
+                });
             };
 
             /**
@@ -82,7 +86,7 @@
                             eventLimit: 4
                         }
                     },
-                    eventClick: self.clickReserva,
+                    eventClick: self.visualizarReserva,
                     dayClick: self.clickDia,
                     eventAfterAllRender: configCalendarioPosRenderizacao,
                     buttonText: {
@@ -106,7 +110,7 @@
                         localId: self.local._id
                     });
 
-                    return ModalService.verReserva(reserva, this.local);
+                    return this.visualizarReserva(reserva);
                 }
             };
 
@@ -197,7 +201,7 @@
              *
              * @return {Promise} Promessa contendo as reservas do local selecionado.
              */
-            this.init = function () {
+            this.carregarReservas = function () {
                 return AgendamentoService.carregarReservasDoLocal(local._id).then(data => {
                     self.reservasSource.splice(0, self.reservasSource.length);
                     const reservas = data.data;
@@ -206,6 +210,6 @@
                 });
             };
 
-            this.init();
+            this.carregarReservas();
         }]);
 })();