+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## unstable
+* Refatoração: Adicionada verificação para o botão "Go Home" ser clicável.
+* Refatoração: Traduzidos textos dos botões de entrar e sair da aplicação.
+* Refatoração: Mudado local para designar título à reserva.
 
 ## 0.2.0
 * Manuteção: Mudado design do modal de reserva.
@@ -13,8 +16,6 @@
 * Funcionalidade: Adcionada função de definir a imagem de capa de um local.
 * Refatoração: A diretiva utilizada para exibição das imagens do local foi trocada por uma mais robusta.
 * Correção: Adicionado botão de voltar em edição e visualização de local e corrigida toolbar.
-* Refatoração: Adicionada verificação para o botão "Go Home" ser clicável.
-* Refatoração: Traduzidos textos dos botões de entrar e sair da aplicação.
 
 ## 0.1.0
 * Funcionalidade: Adicionados tipos "Outro" e "Palestra" para Reserva.